# docs/swagger.yaml
openapi: 3.0.3
info:
    title: ServeBin
    description:
        Welcome to ServeBin documentation! ServeBin is a cutting-edge HTTP
        testing and debugging tool, built with the latest technologies in Go. This documentation
        provides comprehensive details about the endpoints, parameters, and responses
        offered by ServeBin, empowering developers to streamline their testing workflows
        and ensure the reliability of their applications. Explore the various features
        and capabilities of ServeBin to optimize your development process and elevate
        your HTTP testing experience.
    termsOfService: https://servebin.dev/
    contact:
        email: contact@mrayush.me
    license:
        name: BSD 3-Clause
        url: https://github.com/AyushAgnihotri2025/ServeBin/blob/master/LICENSE
    version: v1.3.0
servers:
    - url: https://servebin.dev/
    - url: https://s1.servebin.dev/
    - url: https://s2.servebin.dev/
...
paths:
...
    /image:
        get:
            tags:
                - Images
            summary: Returns image as per the Accept header.
            description: Returns a simple image of the type suggest by the Accept header.
            responses:
                "200":
                    description: OK
                    content:
                        image/png:
                            schema:
                                format: binary
                        image/jpeg:
                            schema:
                                format: binary
                        image/svg+xml:
                            schema:
                                format: binary
                        image/gif:
                            schema:
                                format: binary
                        image/webp:
                            schema:
                                format: binary
                        image/tiff:
                            schema:
                                format: binary
                        image/bmp:
                            schema:
                                format: binary
                        image/apng:
                            schema:
                                format: binary
                        image/avif:
                            schema:
                                format: binary
                        image/x-icon:
                            schema:
                                format: binary
                        image/*:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/apng:
        get:
            tags:
                - Images
            summary: Returns a simple APNG image.
            description: Returns a simple image of the APNG mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/apng:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/avif:
        get:
            tags:
                - Images
            summary: Returns a simple AVIF image.
            description: Returns a simple image of the AVIF mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/avif:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/bmp:
        get:
            tags:
                - Images
            summary: Returns a simple BMP image.
            description: Returns a simple image of the BMP mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/bmp:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/gif:
        get:
            tags:
                - Images
            summary: Returns a simple GIF image.
            description: Returns a simple image of the GIF mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/gif:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/ico:
        get:
            tags:
                - Images
            summary: Returns a simple ICO image.
            description: Returns a simple image of the ICO mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/x-icon:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/jpeg:
        get:
            tags:
                - Images
            summary: Returns a simple JPEG image.
            description: Returns a simple image of the JPEG mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/jpeg:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/png:
        get:
            tags:
                - Images
            summary: Returns a simple PNG image.
            description: Returns a simple image of the PNG mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/png:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/svg:
        get:
            tags:
                - Images
            summary: Returns a simple SVG image.
            description: Returns a simple image of the SVG mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/svg+xml:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/tiff:
        get:
            tags:
                - Images
            summary: Returns a simple TIFF image.
            description: Returns a simple image of the TIFF mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/tiff:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
    /image/webp:
        get:
            tags:
                - Images
            summary: Returns a simple WEBP image.
            description: Returns a simple image of the WEBP mime-type.
            responses:
                "200":
                    description: OK
                    content:
                        image/webp:
                            schema:
                                format: binary
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
...